--- a/resources/test_data/imdb_sample/aka_title.csv
+++ b/resources/test_data/imdb_sample/aka_title.csv
@@ -1,1 +1,1 @@
-562491,0,"""xxxxx"" (None) (None)",NULL,1,NULL,"X25",NULL,NULL,NULL,NULL,"2609c015e7ad09496082f87f6523825b"
+35211,1669200,Yellowcake operazione urano,,1,1978,Y4216,,,,,8c48378c26e30acdf6a9ab3b73280c8e