--- a/resources/test_data/imdb_sample/cast_info.csv.json
+++ b/resources/test_data/imdb_sample/cast_info.csv.json
@@ -1,12 +1,12 @@
 {
-    "config": {"null_handling": "null_string_as_null"},
+    "config": {"null_handling": "null_string_as_value"},
     "columns": [
-        {"name": "id", 			"type": "int",		"nullable": false},
-        {"name": "person_id", 		"type": "int",		"nullable": false},
-        {"name": "movie_id", 		"type": "int",		"nullable": false},
-        {"name": "person_role_id", 	"type": "int", 		"nullable": true},
-        {"name": "note", 		"type": "string", 	"nullable": true},
-        {"name": "nr_order", 		"type": "int", 		"nullable": true},
-        {"name": "role_id", 		"type": "int", 		"nullable": false}
+        {"name": "id",             "type": "int",    "nullable": false},
+        {"name": "person_id",      "type": "int",    "nullable": false},
+        {"name": "movie_id",       "type": "int",    "nullable": false},
+        {"name": "person_role_id", "type": "int",    "nullable": true},
+        {"name": "note",           "type": "string", "nullable": true},
+        {"name": "nr_order",       "type": "int",    "nullable": true},
+        {"name": "role_id",        "type": "int",    "nullable": false}
     ]
 }