--- a/resources/test_data/imdb_sample/char_name.csv
+++ b/resources/test_data/imdb_sample/char_name.csv
@@ -1,1 +1,1 @@
-4301484,"Herself - Cloud Cult Cellist",NULL,NULL,"H6241","C423","6dbb99e1a90e05548dd9ecfb32a236c3"
+24289,President of Handloom Cooperative Society,,,P6235,S23,8fc8d140c869eda53d19ee0983d19705