--- a/resources/test_data/imdb_sample/company_name.csv
+++ b/resources/test_data/imdb_sample/company_name.csv
@@ -1,1 +1,1 @@
-339194,"Century Television Productions",NULL,NULL,"C5363",NULL,"472a24d20a25c4105ac6b98ec4745224"
+45375,Buena Vista International Film Production France,[fr],,B5123,B5123,3d877610178dde125da5afd28f2439a5