--- a/resources/test_data/imdb_sample/company_name.csv.json
+++ b/resources/test_data/imdb_sample/company_name.csv.json
@@ -1,12 +1,12 @@
 {
-    "config": {"null_handling": "null_string_as_null"},
+    "config": {"null_handling": "null_string_as_value"},
     "columns": [
-        {"name": "id", 			"type": "int",		"nullable": false},
-        {"name": "name", 		"type": "string",	"nullable": false},
-        {"name": "country_code", 	"type": "string",	"nullable": true},
-        {"name": "imdb_id", 		"type": "int",		"nullable": true},
-        {"name": "name_pcode_nf", 	"type": "string",	"nullable": true},
-        {"name": "name_pcode_sf", 	"type": "string",	"nullable": true},
-        {"name": "md5sum", 		"type": "string",	"nullable": true}
+        {"name": "id",            "type": "int",    "nullable": false},
+        {"name": "name",          "type": "string", "nullable": false},
+        {"name": "country_code",  "type": "string", "nullable": true},
+        {"name": "imdb_id",       "type": "int",    "nullable": true},
+        {"name": "name_pcode_nf", "type": "string", "nullable": true},
+        {"name": "name_pcode_sf", "type": "string", "nullable": true},
+        {"name": "md5sum",        "type": "string", "nullable": true}
     ]
 }