--- a/resources/test_data/imdb_sample/complete_cast.csv.json
+++ b/resources/test_data/imdb_sample/complete_cast.csv.json
@@ -1,9 +1,9 @@
 {
-    "config": {"null_handling": "null_string_as_null"},
+    "config": {"null_handling": "null_string_as_value"},
     "columns": [
-        {"name": "id", 			"type": "int",		"nullable": false},
-        {"name": "movie_id", 		"type": "int",		"nullable": true},
-        {"name": "subject_id", 		"type": "int",		"nullable": false},
-        {"name": "status_id", 		"type": "int",		"nullable": false}
+        {"name": "id",         "type": "int", "nullable": false},
+        {"name": "movie_id",   "type": "int", "nullable": true},
+        {"name": "subject_id", "type": "int", "nullable": false},
+        {"name": "status_id",  "type": "int", "nullable": false}
     ]
 }