--- a/resources/test_data/imdb_sample/info_type.csv
+++ b/resources/test_data/imdb_sample/info_type.csv
@@ -1,1 +1,1 @@
-113,bottom 10 rank
+17,trivia