--- a/resources/test_data/imdb_sample/keyword.csv
+++ b/resources/test_data/imdb_sample/keyword.csv
@@ -1,1 +1,1 @@
-1136,"sequel","S24"
+4607,extramarital-affair,E2365