--- a/resources/test_data/imdb_sample/keyword.csv.json
+++ b/resources/test_data/imdb_sample/keyword.csv.json
@@ -1,8 +1,8 @@
 {
-    "config": {"null_handling": "null_string_as_null"},
+    "config": {"null_handling": "null_string_as_value"},
     "columns": [
-        {"name": "id", 			"type": "int",		"nullable": false},
-        {"name": "keyword", 		"type": "string",	"nullable": false},
-        {"name": "phonetic_code", 	"type": "string",	"nullable": true}
+        {"name": "id",            "type": "int",    "nullable": false},
+        {"name": "keyword",       "type": "string", "nullable": false},
+        {"name": "phonetic_code", "type": "string", "nullable": true}
     ]
 }