--- a/resources/test_data/imdb_sample/kind_type.csv
+++ b/resources/test_data/imdb_sample/kind_type.csv
@@ -1,1 +1,1 @@
-9,tv short
+7,episode