--- a/resources/test_data/imdb_sample/kind_type.csv.json
+++ b/resources/test_data/imdb_sample/kind_type.csv.json
@@ -1,7 +1,7 @@
 {
-    "config": {"null_handling": "null_string_as_null"},
+    "config": {"null_handling": "null_string_as_value"},
     "columns": [
-        {"name": "id", 			"type": "int",		"nullable": false},
-        {"name": "kind", 		"type": "string",	"nullable": false}
+        {"name": "id",   "type": "int",    "nullable": false},
+        {"name": "kind", "type": "string", "nullable": false}
     ]
 }