--- a/resources/test_data/imdb_sample/link_type.csv
+++ b/resources/test_data/imdb_sample/link_type.csv
@@ -1,1 +1,1 @@
-18,unknown link
+17,alternate language version of