--- a/resources/test_data/imdb_sample/movie_companies.csv.json
+++ b/resources/test_data/imdb_sample/movie_companies.csv.json
@@ -1,10 +1,10 @@
 {
-    "config": {"null_handling": "null_string_as_null"},
+    "config": {"null_handling": "null_string_as_value"},
     "columns": [
-        {"name": "id", 			"type": "int",		"nullable": false},
-        {"name": "movie_id", 		"type": "int",		"nullable": false},
-        {"name": "company_id", 		"type": "int",		"nullable": false},
-        {"name": "company_type_id", 	"type": "int",		"nullable": false},
-        {"name": "note", 		"type": "string",	"nullable": true}
+        {"name": "id",              "type": "int",    "nullable": false},
+        {"name": "movie_id",        "type": "int",    "nullable": false},
+        {"name": "company_id",      "type": "int",    "nullable": false},
+        {"name": "company_type_id", "type": "int",    "nullable": false},
+        {"name": "note",            "type": "string", "nullable": true}
     ]
 }