--- a/resources/test_data/imdb_sample/movie_info_idx.csv.json
+++ b/resources/test_data/imdb_sample/movie_info_idx.csv.json
@@ -1,10 +1,10 @@
 {
-    "config": {"null_handling": "null_string_as_null"},
+    "config": {"null_handling": "null_string_as_value"},
     "columns": [
-        {"name": "id", 			"type": "int",		"nullable": false},
-        {"name": "movie_id", 		"type": "int",		"nullable": false},
-        {"name": "info_type_id", 	"type": "int",		"nullable": false},
-        {"name": "info", 		"type": "string",	"nullable": false},
-        {"name": "note", 		"type": "string",	"nullable": true}
+        {"name": "id",           "type": "int",    "nullable": false},
+        {"name": "movie_id",     "type": "int",    "nullable": false},
+        {"name": "info_type_id", "type": "int",    "nullable": false},
+        {"name": "info",         "type": "string", "nullable": false},
+        {"name": "note",         "type": "string", "nullable": true}
     ]
 }