--- a/resources/test_data/imdb_sample/movie_keyword.csv.json
+++ b/resources/test_data/imdb_sample/movie_keyword.csv.json
@@ -1,8 +1,8 @@
 {
-    "config": {"null_handling": "null_string_as_null"},
+    "config": {"null_handling": "null_string_as_value"},
     "columns": [
-        {"name": "id", 			"type": "int",		"nullable": false},
-        {"name": "movie_id", 		"type": "int",		"nullable": false},
-        {"name": "keyword_id",	 	"type": "int",		"nullable": false}
+        {"name": "id",         "type": "int", "nullable": false},
+        {"name": "movie_id",   "type": "int", "nullable": false},
+        {"name": "keyword_id", "type": "int", "nullable": false}
     ]
 }