--- a/resources/test_data/imdb_sample/movie_link.csv.json
+++ b/resources/test_data/imdb_sample/movie_link.csv.json
@@ -1,9 +1,9 @@
 {
-    "config": {"null_handling": "null_string_as_null"},
+    "config": {"null_handling": "null_string_as_value"},
     "columns": [
-        {"name": "id", 			"type": "int",		"nullable": false},
-        {"name": "movie_id", 		"type": "int",		"nullable": false},
-        {"name": "linked_movie_id",	"type": "int",		"nullable": false},
-        {"name": "link_type_id",	"type": "int",		"nullable": false}
+        {"name": "id",              "type": "int", "nullable": false},
+        {"name": "movie_id",        "type": "int", "nullable": false},
+        {"name": "linked_movie_id", "type": "int", "nullable": false},
+        {"name": "link_type_id",    "type": "int", "nullable": false}
     ]
 }