This PR changes the data set used for the Join Order Benchmark to be the initial data set that has been used for the initial paper ("How Good Are Query Optimizers, Really?". Leis et al. PVLDB Volume 9, No. 3, 2015).
Previously we used the "frozen data set", which is also provided (https://github.com/gregrahn/join-order-benchmark) along with the initial data set and which is more current and larger. 

The problem is that the queries and data do not match as well as with the initial data set. What does that mean? Well, we assume that filters which worked in the initial paper and yielded results are correct. For the updated "frozen" data set, the same filter yields no results. We opened an issue about it (https://github.com/gregrahn/join-order-benchmark/issues/11) but it does not look like the repository is well maintained. *Update:* Viktor Leis confirmed that several queries yield empty results by intention. He also recommended to use the initial data set.

From my point of view, we gain a better reproducility (it's the same data as used in the original paper) and have less surprising empty results. Please note that there are still empty result sets (e.g., query 5a). We believe these are intentional as they happen with both data sets.

The "new" data set is ~2x smaller.

## Data Set Sizes

### Initial data set size
`Martin.Boissier@nemea:~/hyrise_master$ wc -l imdb_data/*.csv`
<details>
<summary>Row counts per file</summary>

```
    901343 aka_name.csv
    361472 aka_title.csv
  36244344 cast_info.csv
   3140339 char_name.csv
         4 comp_cast_type.csv
    234997 company_name.csv
         4 company_type.csv
    135086 complete_cast.csv
       113 info_type.csv
    134170 keyword.csv
         7 kind_type.csv
        18 link_type.csv
   2609129 movie_companies.csv
  14835720 movie_info.csv
   1380035 movie_info_idx.csv
   4523930 movie_keyword.csv
     29997 movie_link.csv
   4167491 name.csv
   2963664 person_info.csv
        12 role_type.csv
   2528312 title.csv
```
</details>

`  74190187 total`

### Updated data set size
`Martin.Boissier@nemea:~/hyrise_master$ wc -l ~/imdb_data/*.csv`
<details>
<summary>Row counts per file</summary>

```
   1312273 aka_name.csv
    528337 aka_title.csv
  63475835 cast_info.csv
   4314863 char_name.csv
         4 comp_cast_type.csv
    362131 company_name.csv
         4 company_type.csv
    135086 complete_cast.csv
       113 info_type.csv
    236627 keyword.csv
         9 kind_type.csv
        18 link_type.csv
   4958296 movie_companies.csv
  29774984 movie_info.csv
   1380035 movie_info_idx.csv
   7480087 movie_keyword.csv
   2585150 movie_link.csv
   6379740 name.csv
   4130207 person_info.csv
        12 role_type.csv
   4736508 title.csv
```
</details>

`  131790319 total`

## Queries with empty results

### With this PR
```
$ grep -c '0 row(s)/0' *PQP.svg | grep -v ":0$"
10b-PQP.svg:5
14b-PQP.svg:1
2c-PQP.svg:6
30b-PQP.svg:1
32a-PQP.svg:11
33a-PQP.svg:1
33c-PQP.svg:1
5a-PQP.svg:11
5b-PQP.svg:12
```

### Current master
```
$ grep -c '0 row(s)/0 ' *PQP.svg | grep -v ":0$"
10b-PQP.svg:12
11b-PQP.svg:5
12b-PQP.svg:20
13a-PQP.svg:19
13b-PQP.svg:23
13c-PQP.svg:23
13d-PQP.svg:19
14a-PQP.svg:3
14b-PQP.svg:10
14c-PQP.svg:1
18b-PQP.svg:8
1b-PQP.svg:10
1d-PQP.svg:10
22a-PQP.svg:1
22b-PQP.svg:10
22c-PQP.svg:1
22d-PQP.svg:1
25b-PQP.svg:31
26a-PQP.svg:27
26b-PQP.svg:25
26c-PQP.svg:32
28a-PQP.svg:1
28b-PQP.svg:1
28c-PQP.svg:1
2c-PQP.svg:9
30a-PQP.svg:30
30b-PQP.svg:44
30c-PQP.svg:2
31a-PQP.svg:38
31b-PQP.svg:48
31c-PQP.svg:24
32a-PQP.svg:20
33a-PQP.svg:4
33b-PQP.svg:3
33c-PQP.svg:1
4b-PQP.svg:5
5a-PQP.svg:14
5b-PQP.svg:15
```


## Performance

With the smaller data set, most queries also run faster. Query 13d is an example for a query that previously had an empty result set and now aggregates ~670,000 rows in the last aggregate. So it's much slower.


<details>
<summary>Configuration Overview - click to expand</summary>

```diff
 +Configuration Overview---------------------------------------------+------------------------------------------------+
 | Parameter        | master_job.json                                | new_data_job.json                              |
 +------------------+------------------------------------------------+------------------------------------------------+
 |  GIT-HASH        | d27fa2304e3607d3cacab90113a821bd361215b3-dirty | d27fa2304e3607d3cacab90113a821bd361215b3-dirty |
 |  benchmark_mode  | Ordered                                        | Ordered                                        |
 |  build_type      | release                                        | release                                        |
 |  chunk_size      | 65535                                          | 65535                                          |
 |  clients         | 1                                              | 1                                              |
 |  compiler        | gcc 10.3                                       | gcc 10.3                                       |
 |  cores           | 0                                              | 0                                              |
 |  date            | 2021-06-24 13:14:37                            | 2021-06-24 18:04:59                            |
 |  encoding        | {'default': {'encoding': 'Dictionary'}}        | {'default': {'encoding': 'Dictionary'}}        |
 |  indexes         | False                                          | False                                          |
 |  max_duration    | 60000000000                                    | 60000000000                                    |
 |  max_runs        | -1                                             | -1                                             |
 |  time_unit       | ns                                             | ns                                             |
 |  using_scheduler | False                                          | False                                          |
 |  verify          | False                                          | False                                          |
 |  warmup_duration | 0                                              | 0                                              |
 +------------------+------------------------------------------------+------------------------------------------------+
```
</details>

```diff
 +---------++----------+---------+--------++----------+----------+----------+---------+
 | Item    || Latency (ms/iter)  | Change || Throughput (iter/s) |   Change | p-value |
 |         ||      old |     new |        ||      old |      new |          |         |
 +---------++----------+---------+--------++----------+----------+----------+---------+
+| 10a     ||    319.3 |   171.9 |  -46%  ||     3.13 |     5.82 |    +86%  |  0.0000 |
+| 10b     ||    313.1 |   123.7 |  -60%  ||     3.19 |     8.08 |   +153%  |  0.0000 |
+| 10c     ||    732.0 |   350.4 |  -52%  ||     1.37 |     2.85 |   +109%  |  0.0000 |
+| 11a     ||    129.6 |    50.9 |  -61%  ||     7.71 |    19.65 |   +155%  |  0.0000 |
+| 11b     ||    122.7 |    50.8 |  -59%  ||     8.15 |    19.69 |   +142%  |  0.0000 |
+| 11c     ||    466.2 |    48.0 |  -90%  ||     2.14 |    20.84 |   +872%  |  0.0000 |
+| 11d     ||   4713.4 |    27.4 |  -99%  ||     0.21 |    36.56 | +17131%  |  0.0000 |
+| 12a     ||    256.6 |    51.8 |  -80%  ||     3.90 |    19.29 |   +395%  |  0.0000 |
+| 12b     ||    284.0 |   138.0 |  -51%  ||     3.52 |     7.25 |   +106%  |  0.0000 |
+| 12c     ||   3323.2 |  1590.7 |  -52%  ||     0.30 |     0.63 |   +109%  |  0.0000 |
-| 13a     ||    229.2 |   356.6 |  +56%  ||     4.36 |     2.80 |    -36%  |  0.0000 |
+| 13b     ||    343.2 |   176.4 |  -49%  ||     2.91 |     5.67 |    +95%  |  0.0000 |
+| 13c     ||    211.5 |   108.8 |  -49%  ||     4.73 |     9.19 |    +94%  |  0.0000 |
-| 13d     ||    228.9 |   702.0 | +207%  ||     4.37 |     1.42 |    -67%  |  0.0000 |
+| 14a     ||   3820.2 |  2337.2 |  -39%  ||     0.26 |     0.43 |    +63%  |  0.0000 |
+| 14b     ||   4176.1 |  2271.3 |  -46%  ||     0.24 |     0.44 |    +84%  |  0.0000 |
+| 14c     ||   3817.5 |  2426.7 |  -36%  ||     0.26 |     0.41 |    +57%  |  0.0000 |
+| 15a     ||    186.7 |    86.4 |  -54%  ||     5.36 |    11.58 |   +116%  |  0.0000 |
+| 15b     ||    183.3 |    85.6 |  -53%  ||     5.46 |    11.68 |   +114%  |  0.0000 |
+| 15c     ||    146.3 |    50.1 |  -66%  ||     6.84 |    19.95 |   +192%  |  0.0000 |
+| 15d     ||    374.9 |    87.9 |  -77%  ||     2.67 |    11.38 |   +327%  |  0.0000 |
+| 16a     ||   4267.9 |  2440.7 |  -43%  ||     0.23 |     0.41 |    +75%  |  0.0000 |
+| 16b     ||   5796.5 |  3536.5 |  -39%  ||     0.17 |     0.28 |    +64%  |  0.0000 |
+| 16c     ||   4426.6 |  2548.5 |  -42%  ||     0.23 |     0.39 |    +74%  |  0.0000 |
+| 16d     ||   4373.9 |  2507.8 |  -43%  ||     0.23 |     0.40 |    +74%  |  0.0000 |
+| 17a     ||    857.6 |   575.3 |  -33%  ||     1.17 |     1.74 |    +49%  |  0.0000 |
+| 17b     ||    684.9 |   455.6 |  -33%  ||     1.46 |     2.19 |    +50%  |  0.0000 |
+| 17c     ||    652.2 |   433.1 |  -34%  ||     1.53 |     2.31 |    +51%  |  0.0000 |
+| 17d     ||    761.5 |   512.2 |  -33%  ||     1.31 |     1.95 |    +49%  |  0.0000 |
+| 17e     ||   2440.5 |  1598.5 |  -35%  ||     0.41 |     0.63 |    +53%  |  0.0000 |
+| 17f     ||   1399.1 |   958.9 |  -31%  ||     0.71 |     1.04 |    +46%  |  0.0000 |
-| 18a     ||    799.5 |  1674.5 | +109%  ||     1.25 |     0.60 |    -52%  |  0.0000 |
+| 18b     ||    230.2 |   133.0 |  -42%  ||     4.34 |     7.52 |    +73%  |  0.0000 |
+| 18c     ||   3429.9 |  2179.0 |  -36%  ||     0.29 |     0.46 |    +57%  |  0.0000 |
+| 19a     ||   5951.5 |  3205.8 |  -46%  ||     0.17 |     0.31 |    +86%  |  0.0000 |
+| 19b     ||   3032.8 |  1604.3 |  -47%  ||     0.33 |     0.62 |    +89%  |  0.0000 |
+| 19c     ||   5773.2 |  3083.6 |  -47%  ||     0.17 |     0.32 |    +87%  |  0.0000 |
+| 19d     ||   4886.5 |  2234.4 |  -54%  ||     0.20 |     0.45 |   +119%  |  0.0000 |
+| 1a      ||     71.2 |    33.4 |  -53%  ||    14.05 |    29.91 |   +113%  |  0.0000 |
+| 1b      ||     20.2 |    10.4 |  -48%  ||    49.50 |    95.95 |    +94%  |  0.0000 |
+| 1c      ||     20.2 |    11.4 |  -44%  ||    49.50 |    87.92 |    +78%  |  0.0000 |
+| 1d      ||     20.0 |    10.4 |  -48%  ||    50.00 |    96.00 |    +92%  |  0.0000 |
+| 20a     ||   1216.4 |   775.3 |  -36%  ||     0.82 |     1.29 |    +57%  |  0.0000 |
+| 20b     ||   1072.5 |   669.0 |  -38%  ||     0.93 |     1.49 |    +60%  |  0.0000 |
+| 20c     ||    724.0 |   484.3 |  -33%  ||     1.38 |     2.06 |    +49%  |  0.0000 |
+| 21a     ||   3490.0 |    77.1 |  -98%  ||     0.29 |    12.98 |  +4429%  |  0.0000 |
+| 21b     ||    233.5 |    56.9 |  -76%  ||     4.28 |    17.56 |   +310%  |  0.0000 |
+| 21c     ||   3647.6 |    77.1 |  -98%  ||     0.27 |    12.96 |  +4628%  |  0.0000 |
+| 22a     ||   3100.9 |  1766.4 |  -43%  ||     0.32 |     0.57 |    +76%  |  0.0000 |
+| 22b     ||   3102.3 |  1661.1 |  -46%  ||     0.32 |     0.60 |    +87%  |  0.0000 |
+| 22c     ||   3788.6 |  2666.9 |  -30%  ||     0.26 |     0.37 |    +42%  |  0.0000 |
+| 22d     ||   3814.8 |  2758.9 |  -28%  ||     0.26 |     0.36 |    +38%  |  0.0000 |
+| 23a     ||    143.8 |    56.3 |  -61%  ||     6.96 |    17.76 |   +155%  |  0.0000 |
+| 23b     ||     98.1 |    53.7 |  -45%  ||    10.19 |    18.61 |    +83%  |  0.0000 |
+| 23c     ||    160.2 |    61.7 |  -61%  ||     6.24 |    16.20 |   +160%  |  0.0000 |
+| 24a     ||   3014.1 |  1581.7 |  -48%  ||     0.33 |     0.63 |    +91%  |  0.0000 |
+| 24b     ||   2957.7 |  1545.6 |  -48%  ||     0.34 |     0.65 |    +91%  |  0.0000 |
-| 25a     ||    208.6 |   291.1 |  +40%  ||     4.79 |     3.44 |    -28%  |  0.0000 |
+| 25b     ||    227.3 |   132.4 |  -42%  ||     4.40 |     7.55 |    +72%  |  0.0000 |
+| 25c     ||   3438.0 |  2581.5 |  -25%  ||     0.29 |     0.39 |    +33%  |  0.0000 |
+| 26a     ||    623.0 |   441.5 |  -29%  ||     1.61 |     2.26 |    +41%  |  0.0000 |
+| 26b     ||    615.0 |   405.5 |  -34%  ||     1.63 |     2.47 |    +52%  |  0.0000 |
+| 26c     ||    623.8 |   500.0 |  -20%  ||     1.60 |     2.00 |    +25%  |  0.0000 |
+| 27a     ||   3135.3 |  1575.0 |  -50%  ||     0.32 |     0.63 |    +99%  |  0.0000 |
+| 27b     ||   3112.1 |  1574.1 |  -49%  ||     0.32 |     0.64 |    +98%  |  0.0000 |
+| 27c     ||    178.6 |    77.3 |  -57%  ||     5.60 |    12.93 |   +131%  |  0.0000 |
+| 28a     ||   3842.1 |  2286.1 |  -40%  ||     0.26 |     0.44 |    +68%  |  0.0000 |
+| 28b     ||   3073.9 |  1569.1 |  -49%  ||     0.33 |     0.64 |    +96%  |  0.0000 |
+| 28c     ||   3833.6 |  2361.2 |  -38%  ||     0.26 |     0.42 |    +62%  |  0.0000 |
+| 29a     ||   2968.8 |  1619.5 |  -45%  ||     0.34 |     0.62 |    +83%  |  0.0000 |
+| 29b     ||    177.1 |   105.1 |  -41%  ||     5.65 |     9.51 |    +68%  |  0.0000 |
+| 29c     ||   3174.8 |  1619.5 |  -49%  ||     0.31 |     0.62 |    +96%  |  0.0000 |
+| 2a      ||     82.1 |    48.5 |  -41%  ||    12.18 |    20.60 |    +69%  |  0.0000 |
+| 2b      ||     66.2 |    40.3 |  -39%  ||    15.10 |    24.83 |    +64%  |  0.0000 |
+| 2c      ||     53.5 |    31.8 |  -40%  ||    18.71 |    31.43 |    +68%  |  0.0000 |
+| 2d      ||    106.0 |    67.8 |  -36%  ||     9.44 |    14.75 |    +56%  |  0.0000 |
+| 30a     ||    221.5 |   164.4 |  -26%  ||     4.51 |     6.08 |    +35%  |  0.0000 |
+| 30b     ||    560.4 |   294.8 |  -47%  ||     1.78 |     3.39 |    +90%  |  0.0000 |
+| 30c     ||   3467.9 |  2015.2 |  -42%  ||     0.29 |     0.50 |    +72%  |  0.0000 |
+| 31a     ||    259.2 |   169.3 |  -35%  ||     3.86 |     5.91 |    +53%  |  0.0000 |
+| 31b     ||    593.2 |   312.1 |  -47%  ||     1.69 |     3.20 |    +90%  |  0.0000 |
+| 31c     ||    262.8 |   193.3 |  -26%  ||     3.80 |     5.17 |    +36%  |  0.0000 |
+| 32a     ||     33.9 |    15.7 |  -54%  ||    29.51 |    63.52 |   +115%  |  0.0000 |
+| 32b     ||    251.8 |    34.0 |  -87%  ||     3.97 |    29.43 |   +641%  |  0.0000 |
+| 33a     ||     61.2 |    24.8 |  -60%  ||    16.34 |    40.39 |   +147%  |  0.0000 |
+| 33b     ||     60.1 |    24.1 |  -60%  ||    16.63 |    41.56 |   +150%  |  0.0000 |
+| 33c     ||     70.9 |    29.5 |  -58%  ||    14.10 |    33.91 |   +140%  |  0.0000 |
+| 3a      ||   3541.7 |  1824.7 |  -48%  ||     0.28 |     0.55 |    +94%  |  0.0000 |
+| 3b      ||     32.0 |    16.6 |  -48%  ||    31.22 |    60.15 |    +93%  |  0.0000 |
+| 3c      ||   4420.4 |  2302.8 |  -48%  ||     0.23 |     0.43 |    +92%  |  0.0000 |
-| 4a      ||    101.7 |   261.0 | +157%  ||     9.83 |     3.83 |    -61%  |  0.0000 |
+| 4b      ||     24.5 |    15.0 |  -39%  ||    40.88 |    66.69 |    +63%  |  0.0000 |
-| 4c      ||    111.4 |   295.2 | +165%  ||     8.98 |     3.39 |    -62%  |  0.0000 |
+| 5a      ||   3432.2 |  1759.0 |  -49%  ||     0.29 |     0.57 |    +95%  |  0.0000 |
+| 5b      ||    261.9 |   144.7 |  -45%  ||     3.82 |     6.91 |    +81%  |  0.0000 |
+| 5c      ||   4078.5 |  2074.5 |  -49%  ||     0.25 |     0.48 |    +97%  |  0.0000 |
+| 6a      ||    203.3 |   114.9 |  -43%  ||     4.92 |     8.70 |    +77%  |  0.0000 |
+| 6b      ||    884.2 |   127.5 |  -86%  ||     1.13 |     7.84 |   +593%  |  0.0000 |
+| 6c      ||    203.2 |   109.8 |  -46%  ||     4.92 |     9.10 |    +85%  |  0.0000 |
+| 6d      ||    884.5 |   522.3 |  -41%  ||     1.13 |     1.91 |    +69%  |  0.0000 |
+| 6e      ||    203.1 |   115.4 |  -43%  ||     4.92 |     8.67 |    +76%  |  0.0000 |
+| 6f      ||   1416.5 |   723.5 |  -49%  ||     0.71 |     1.38 |    +96%  |  0.0000 |
+| 7a      ||    208.4 |    93.6 |  -55%  ||     4.80 |    10.69 |   +123%  |  0.0000 |
+| 7b      ||    138.2 |    71.5 |  -48%  ||     7.23 |    13.99 |    +93%  |  0.0000 |
+| 7c      ||  10508.8 |   627.0 |  -94%  ||     0.10 |     1.59 |  +1576%  |       ˅ |
+| 8a      ||    103.3 |    56.4 |  -45%  ||     9.68 |    17.73 |    +83%  |  0.0000 |
+| 8b      ||    161.2 |    93.4 |  -42%  ||     6.20 |    10.71 |    +73%  |  0.0000 |
+| 8c      ||   3416.2 |  1586.5 |  -54%  ||     0.29 |     0.63 |   +115%  |  0.0000 |
+| 8d      ||    970.9 |   467.8 |  -52%  ||     1.03 |     2.14 |   +108%  |  0.0000 |
+| 9a      ||   3234.4 |  1779.3 |  -45%  ||     0.31 |     0.56 |    +82%  |  0.0000 |
+| 9b      ||    321.5 |   148.8 |  -54%  ||     3.11 |     6.72 |   +116%  |  0.0000 |
+| 9c      ||   3260.7 |  1744.7 |  -46%  ||     0.31 |     0.57 |    +87%  |  0.0000 |
+| 9d      ||   3876.2 |  1950.8 |  -50%  ||     0.26 |     0.51 |    +99%  |  0.0000 |
 +---------++----------+---------+--------++----------+----------+----------+---------+
+| Sum     || 186220.0 | 94963.0 |  -49%  ||          |          |          |         |
+| Geomean ||          |         |        ||          |          |   +104%  |         |
 +---------++----------+---------+--------++----------+----------+----------+---------+
 |         || ˅ Insufficient number of runs for p-value calculation                   |
 +---------++----------+---------+--------++----------+----------+----------+---------+
```

# Notes on Join Order Benchmark data

The paper that introduced the Join Order Benchmark ("How Good Are Query Optimizers, Really?". Viktor Leis et al., PVLDB Volume 9, No. 3, 2015) provided information about two datasets:
 - the original data set: this data set from 2013 is used in the paper and can be downloaded [here](http://homepages.cwi.nl/~boncz/job/imdb.tgz).
 - the "frozen data set" (provided by the Freie Universität Berlin): this is a more recent dump of the IMDB data; the authors of the paper provided intructions for generating a relational data set using this dump

Hyrise used the frozen data set until July 2021 (#2376), but due to issues with empty queries the used data set has been changed to the original data set used in the paper (see https://github.com/hyrise/hyrise/pull/2376).
As the CSV files of the paper data set do not follow RFC 4180 (i.e., using "" to escape an " in a string, which Hyrise assumes to be the case for CSV files), the following command can be used to adapt the CSV files (tested with MacOS 11.4):
`for csv_file in *.csv; do echo $csv_file; sed -i'' -e 's/\\\\\\"/3BS_MARKER/g;s/\\\\"/2BS_MARKER2/g;s/\\"/""/g;s/3BS_MARKER/\\""/g;s/2BS_MARKER2/\\"/g' $csv_file; done`

The command is a rather hard-coded fix to avoid using lookbehind regexes. Please be cautious using it on a more recent export of the Join Order Benchmark data.

Exemplary lines before adaption:
- movie_info.csv:	`127472,2130098,13,"FACT: Dunn uploads a file from an Apple Powerbook in \"C:\\\", which would be appropriate for a DOS/Windows system.",`
- person_info.csv:	`2010004,1163574,25,"CD: \"All-Time Hits, Vol. 2\"\\",`

Reformatted:
- movie_info.csv:	`127472,2130098,13,"FACT: Dunn uploads a file from an Apple Powerbook in ""C:\"", which would be appropriate for a DOS/Windows system.",`
- person_info.csv:	`2010004,1163574,25,"CD: ""All-Time Hits, Vol. 2""\",`


## Reproducibility: previously used data set

In case you need the previous frozen data set for reproducibility reasons, you can find it [here](http://www.dropbox.com/s/fgju7cifpehqlww/imdb.zip?dl=1).


## Notes on the initial data set

…