@Bouncner 2021-06-25
I'd like to get opinions on this PR before proceeding (especially pinging @Bensk1 and @julianmenzler).

As soon as we decide to continue, I'll update the data set in the PR (it's just a place holder right now).