@Bensk1 2021-06-25
First, tremendous job improving the performance of so many queries! I think, we should use the initial dataset to produce the _intended_ query results. However, since we have used the broken frozen dataset for published papers and master's theses, would it make sense to implement a switch that utilizes the frozen dataset on demand to enable reproducibility of old results? By default, we would always use the initial (correct) version.

As a side note: due to the much smaller size of the new data set, it should be storable with Git Large File Storage (limit 2GB) now.