@Bouncner 2021-06-25
As far as I can see, we don't have a paper with JOB data (yet). So from my side a printed comment that one should be aware of the two data sets and that we switched should be sufficient.

The git lfs usage would be definitely nice.