@Bensk1 2021-06-25
> As far as I can see, we don't have a paper with JOB data (yet). So from my side a printed comment that one should be aware of the two data sets and that we switched should be sufficient.

Sounds good to me. I would still vote for keeping the old dataset where it is and maybe add a short manual to the Wiki to explain how to use it.