@Bouncner 2021-06-29
> As a side note: due to the much smaller size of the new data set, it should be storable with Git Large File Storage (limit 2GB) now

I was thinking about how to do that.  Putting the zip on LFS doesn't sound too good to me as we don't need to keep it around. I haven't find a way to remove LFS files after loading them once. We could ship binary table data via LFS, but these are too unstable over time I think. CSV files are unnecessarily large to have them in LFS.

And more over, deleting files from LFS still appears to be a major pain even in the year 2021.

So right now with my limited experience with LFS, I tend to use something else.