@Bensk1 2021-06-30
Okay, let's stick with the current storage solution.