@Bouncner 2021-07-02
I would add the following files as a `notes.md` to the ZIP file. Any comments?

================

### Notes on Join Order Benchmark data

We use the CSV files of the initial paper ("How Good Are Query Optimizers, Really?". Viktor Leis et al., PVLDB Volume 9, No. 3, 2015): http://homepages.cwi.nl/~boncz/job/imdb.tgz.
We have previously used an updated data set ("frozen data set" provided by the FU Berlin with the provided instructions of Viktor Leis) but found issues with several queries yielding no results.

As the provided CSV files do not follow RFC 4180 (i.e., using "" to escape an " in a string, which we assume to be the case for CSV files), we used the following command to adapt the CSV files (tested with MacOS 11.4):
`for csv_file in *.csv; do echo $csv_file; sed -i'' -e 's/\\\\\\"/3BS_MARKER/g;s/\\\\"/2BS_MARKER2/g;s/\\"/""/g;s/3BS_MARKER/\\""/g;s/2BS_MARKER2/\\"/g' $csv_file; done`

The command is a rather hard-coded fix to avoid using lookbehind regexes. Please be cautious using it on a more recent export of the Join Order Benchmark data.

Exemplary lines before adaption:
- movie_info.csv:	`127472,2130098,13,"FACT: Dunn uploads a file from an Apple Powerbook in \"C:\\\", which would be appropriate for a DOS/Windows system.",`
- person_info.csv:	`2010004,1163574,25,"CD: \"All-Time Hits, Vol. 2\"\\",`

Reformatted:
- movie_info.csv:	`127472,2130098,13,"FACT: Dunn uploads a file from an Apple Powerbook in ""C:\"", which would be appropriate for a DOS/Windows system.",`
- person_info.csv:	`2010004,1163574,25,"CD: ""All-Time Hits, Vol. 2""\",`