@Bensk1 2021-07-02
Sounds good to me. Also, we should think about incorporating: https://homepages.cwi.nl/~boncz/job/readme.txt somehow