@Bouncner 2021-07-05
FYI: I emailed the author of the Join Order Benchmark repository and asked if we could add some further information (not maintained, please use the initial data set, etc.) in the readme.