@Bensk1 2021-07-05
Also, do we need to adapt the test scripts after changing `resources/test_data/imdb_sample/*`?