@Bouncner 2021-07-05
> Merge master to remove change in third_party/benchmark from change log?

Done.

> Also, do we need to adapt the test scripts after changing `resources/test_data/imdb_sample/*`?

Not that I am aware of. Locally it ran just fine.